requirements.txt
python-binance==1.0.19
pandas==2.2.3
numpy==2.1.3
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
streamlit==1.38.0
hyperliquid-python-sdk>=0.9.0
eth-account>=0.10.0
ccxt
streamlit-autorefresh
cryptography
